Refactor redundant code and fix clippy warnings
- Renamed `ParsedDesignatedInitializerDesignator` to `ParsedDesignator` and removed redundant alias.
- Encapsulated `TypeRef` dummy initialization in `TypeRef::dummy()`.
- Fixed multiple clippy warnings:
  - Removed unnecessary casts and boolean logic.
  - Simplified string handling (`strip_suffix`, avoid extra slicing).
  - Collapsed nested `if` statements and `match` arms.
  - Removed needless `else` blocks.

+++ src/semantic/lower_initializer.rs
@@ -460,7 +460,7 @@ impl<'a> AstToMirLowerer<'a> {
         }
 
         let align = elem_layout.alignment;
-        let stride = elem_layout.size as u16;
+        let stride = elem_layout.size;
         // let total_size = (stride as usize * size) as u16;
 
         let array_ty = self.mir_builder.add_type(MirType::Array {
